xs2a-impl/src/main/java/de/adorsys/psd2/xs2a/web/mapper/HrefLinkMapper.java
/*
 * Copyright 2018-2022 adorsys GmbH & Co KG
 *
 * This program is free software: you can redistribute it and/or modify it
 * under the terms of the GNU Affero General Public License as published
 * by the Free Software Foundation, either version 3 of the License, or (at
 * your option) any later version. This program is distributed in the hope that
 * it will be useful, but WITHOUT ANY WARRANTY; without even the implied
 * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 * See the GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program. If not, see https://www.gnu.org/licenses/.
 *
 * This project is also available under a separate commercial license. You can
 * contact us at psd2@adorsys.com.
 */

package de.adorsys.psd2.xs2a.web.mapper;

import com.fasterxml.jackson.core.type.TypeReference;
import de.adorsys.psd2.mapper.Xs2aObjectMapper;
import de.adorsys.psd2.xs2a.domain.HrefType;
import de.adorsys.psd2.xs2a.domain.Links;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Component;

import java.util.Map;

@Component
@RequiredArgsConstructor
public class HrefLinkMapper {
    private final Xs2aObjectMapper xs2aObjectMapper;

    /**
     * Maps Links to Links Map
     *
     * @param links Links model Object, where URI links stored as a HrefType
     * @return Map with link name and HrefType value.
     */
    public Map<String, HrefType> mapToLinksMap(Links links) {
        if (links == null) {
            return null;
        }

        return xs2aObjectMapper.convertValue(links, new TypeReference<Map<String, HrefType>>() {
        });
    }
}
